pyproject.toml
[tool.poetry]
name = "tests"
version = "0.1.0"
description = ""
authors = ["Deni <denibatyzhev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.14.0"
django = "3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
